+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [Development]
+- Signals: added `ticker_universe` to get the list of accepted tickers
+
 ## [2.3.1] - 2020-05-06 - "Signals"
 - fix Signals submission upload (#25)